Self Generated

+++ .profile.sh
@@ -21,5 +21,5 @@ alias revertTmp="[[ -f $DIR/backup/.temp_aliases.sh ]] && cat $DIR/backup/.temp_
 ########## Set-up git filter-repo ##########
 export PATH="${HOME}/bin:${PATH}"
 mkdir -p ~/bin
-curl https://raw.githubusercontent.com/newren/git-filter-repo/main/git-filter-repo -o ~/bin/git-filter-repo 
+[[ ! -f "~/bin/git-filter-repo" ]] || curl https://raw.githubusercontent.com/newren/git-filter-repo/main/git-filter-repo -o ~/bin/git-filter-repo 
 chmod +x ~/bin/git-filter-repo
